backend/requirements-full.txt
# requirements.txt para Render (Django + PostgreSQL)

asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
dj-database-url==3.0.1
Django==5.0.14
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
mccabe==0.7.0
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.11.0


# Solo para desarrollo (no instalar en producción)
black==25.9.0
django-debug-toolbar==6.1.0
django-extensions==4.1
factory_boy==3.3.3
Faker==37.12.0
flake8==7.3.0
isort==7.0.0
mypy_extensions==1.1.0
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
pytokens==0.2.0
